chore(i18n): remove legacy key + rename
