+++ types/config.ts
@@ -13,17 +13,6 @@ export interface OGConfig {
   imgHeight: string;
 }
 
-export interface FooterConfig {
-  text: string;
-  link: string;
-}
-
-export interface SocialConfig {
-  icon: string;
-  link: string;
-  alt?: string;
-}
-
 export interface SiteConfig {
   title: string;
   description: string;
@@ -35,6 +24,4 @@ export interface SiteConfig {
   rssFeeds: Array<RSSFeed>;
   websiteBanners: Record<string, WebsiteBanner>;
   websiteBadges: Record<string, WebsiteBadge>;
-  footerLinks: Array<FooterConfig>;
-  socialLinks: Array<SocialConfig>;
 }
